Add settings, add not-so-bad breakout/brick breaker
This commit introduces game.settings, which will be used by the
settings menu in the future.

It also adds breakout.lua, which is a pretty alright breakout
implementation. Because of my stubborn insistence on being pixel-
perfect, the collision is not SUPER interesting, however it is
much better than the original game. However to make the separation
between bricks more obvious I do shrink them slightly, so maybe
this commitment to pixel-perfect drawing is stupid after all.